KOGITO-1901: Rename DE column picker
